revert

+++ ydb/tests/olap/ttl_tiering/ttl_delete_s3.py
@@ -267,7 +267,8 @@ class TestDeleteS3Ttl(TllDeleteBase):
         logger.info(f"Rows older than {self.days_to_cool} days: {self.get_row_count_by_date(table_path, self.days_to_cool)}")
         logger.info(f"Rows older than {self.days_to_freeze} days: {self.get_row_count_by_date(table_path, self.days_to_freeze)}")
 
-        assert ColumnTableHelper.portions_actualized_in_sys(self.table)
+        if not self.wait_for(lambda: self.portions_actualized_in_sys(self.table), 120):
+            raise Exception(".sys reports incorrect data portions")
 
         stmt = f"""
             DELETE FROM `{table_path}`
